requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1
astroid==2.4.2; python_version >= '3.5'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
coverage==5.4
django-filter==2.4.0
django==3.1.6
djangorestframework==3.12.2
iniconfig==1.1.1
isort==5.7.0; python_version >= '3.6' and python_version < '4.0'
lazy-object-proxy==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown==3.3.3
mccabe==0.6.1
nodeenv==1.5.0
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pylint-django==2.4.2
pylint-plugin-utils==0.6
pylint==2.6.2; python_version >= '3.5'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-django==4.1.0
pytest==6.2.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
wrapt==1.12.1
